refactor: 움직일 수 있는 최소 단위를 반환하는 함수명 변경
