+++ src/main/java/chess/domain/Calculator.java
@@ -1,7 +1,7 @@
-package chess;
+package chess.domain;
 
 public class Calculator {
-    public static int calculateSign(int value) {
+    public static int calculateMinMovement(int value) {
         if (value == 0) {
             return 0;
         }
